feat: MeetingVote에서 footer 분리
