+++ src/pages/MeetingVote/MeetingVote.tsx
@@ -1,44 +1,37 @@
 import Box from '@mui/material/Box';
-import Button from '@mui/material/Button';
 import Typography from '@mui/material/Typography';
-import { useMutation } from '@tanstack/react-query';
-import { AxiosError } from 'axios';
-import { UIEvent, useEffect, useRef, useState } from 'react';
+import { UIEvent, useEffect, useEffect, useRef, useState } from 'react';
 import { useResizeDetector } from 'react-resize-detector';
 import { useNavigate, useSearchParams } from 'react-router-dom';
 import { useRecoilState, useRecoilValue, useResetRecoilState, useSetRecoilState } from 'recoil';
 
-import { createVoting, updateVoting } from '../../apis/votes';
 import WritingHands from '../../assets/writing.svg';
 import { ScrollDownFloatingButton } from '../../components/buttons/ScrollDownFloatingButton';
 import { Loading } from '../../components/Loading';
-import { Contents, Footer, Header, HeaderContainer, Page } from '../../components/pageLayout';
-import { FlexVertical, FullHeightButtonGroup } from '../../components/styled';
+import { Contents, Header, HeaderContainer, Page } from '../../components/pageLayout';
+import { FlexVertical } from '../../components/styled';
 import { UserList, UserListData } from '../../components/UserList/UserList';
 import { VoteTable } from '../../components/VoteTable/VoteTable';
 import { MeetingType } from '../../constants/meeting';
 import { useMeeting } from '../../hooks/useMeeting';
 import { useMeetingViewVoteMode } from '../../hooks/useMeetingVote';
-import { useProgress } from '../../hooks/useProgress';
 import { useVotings } from '../../hooks/useVotings';
 import { currentUserStateFamily } from '../../stores/currentUser';
 import { currentUserVotingSlotsState } from '../../stores/currentUserVotingSlots';
-import { showVoteSuccessPopupState } from '../../stores/showVoteSuccessPopup';
 import { userListState, votingsState as votingRecoilState } from '../../stores/voting';
-import { InputUsernameModal } from '../../templates/MeetingView/InputUsernameModal';
 import { PrimaryBold, VoteTableWrapper } from '../../templates/MeetingView/styled';
+import MeetingVoteFooter from '../../templates/MeetingVote/MeetingVoteFooter';
 
 function MeetingVote() {
   const [searchParams] = useSearchParams();
   const pageElementRef = useRef<HTMLDivElement>(null);
   const { meeting, meetingId, isFetching: isMeetingFetching } = useMeeting();
-  const { votings, isFetching: isVotingsFetcing, invalidateVotings } = useVotings();
+  const { votings, isFetching: isVotingsFetcing } = useVotings();
   const isFetching = isMeetingFetching && isVotingsFetcing;
 
   const [currentUser, setCurrentUser] = useRecoilState(currentUserStateFamily(meetingId));
   const resetCurrentUser = useResetRecoilState(currentUserStateFamily(meetingId));
   const setCurrentUserVotingSlots = useSetRecoilState(currentUserVotingSlotsState);
-  const setShowVoteSuccessPopup = useSetRecoilState(showVoteSuccessPopupState);
   const isNewUser = !currentUser;
 
   const [votingsState, setVotingsState] = useRecoilState(votingRecoilState);
@@ -48,62 +41,12 @@ function MeetingVote() {
     checked: user.id === currentUser?.id,
   }));
 
-  const [showUsernameModal, setShowUsernameModal] = useState<boolean>(false);
   const [hasMoreBottomScroll, setHasMoreBottomScroll] = useState<boolean>(false);
 
   const navigate = useNavigate();
-  const { show, hide } = useProgress();
 
-  const { mutate: updateVotingMutate } = useMutation({
-    mutationFn: updateVoting,
-    onMutate: () => show(),
-    onSuccess: async () => {
-      setShowUsernameModal(false);
-      setShowVoteSuccessPopup(true);
-      await invalidateVotings();
-      navigate(`/meetings/${meetingId}`);
-    },
-    onError: (error) => {
-      if (error instanceof AxiosError) {
-        const errorResponseData = error.response?.data as { message: string };
-        alert(errorResponseData?.message);
-      } else {
-        alert(error);
-      }
-    },
-    onSettled: () => hide(),
-  });
-
-  const { mutate: createVotingMutate } = useMutation({
-    mutationFn: createVoting,
-    onMutate: () => show(),
-    onSuccess: async (res) => {
-      setCurrentUser({
-        id: res.id,
-        username: res.username,
-      });
-      setShowUsernameModal(false);
-      setShowVoteSuccessPopup(true);
-      await invalidateVotings();
-      navigate(`/meetings/${meetingId}`);
-    },
-    onError: (error) => {
-      if (error instanceof AxiosError) {
-        const errorResponseData = error.response?.data as { message: string };
-        alert(errorResponseData?.message);
-      } else {
-        alert(error);
-      }
-    },
-    onSettled: () => hide(),
-  });
-
-  const {
-    voteTableDataList,
-    currentUserVotingSlots,
-    handleClickVoteTableSlot,
-    handleClickVoteTableDate,
-  } = useMeetingViewVoteMode(meeting);
+  const { voteTableDataList, handleClickVoteTableSlot, handleClickVoteTableDate } =
+    useMeetingViewVoteMode(meeting);
 
   useEffect(() => {
     const isFromSharedURL = searchParams.get('ref') === 'share';
@@ -153,42 +96,6 @@ function MeetingVote() {
     const previousVoting = votingsState.find((voting) => voting.username === clickedUser.username);
     const previousVotingSlots = previousVoting?.[meeting.type];
     setCurrentUserVotingSlots(previousVotingSlots ?? []);
-  };
-
-  // eslint-disable-next-line @typescript-eslint/require-await
-  const handleClickVote = async () => {
-    if (!meeting) {
-      return;
-    }
-
-    if (isNewUser) {
-      setShowUsernameModal(true);
-      return;
-    }
-
-    // Old user
-    updateVotingMutate({
-      meetingId,
-      votingId: currentUser.id,
-      data: {
-        username: currentUser.username,
-        [meeting.type]: currentUserVotingSlots,
-      },
-    });
-  };
-
-  const handleUsernameConfirm = (username: string) => {
-    if (!meeting) {
-      return;
-    }
-
-    createVotingMutate({
-      meetingId,
-      data: {
-        username,
-        [meeting.type]: currentUserVotingSlots,
-      },
-    });
   };
 
   const handlePageScroll = (event: UIEvent<HTMLDivElement>) => {
@@ -277,37 +184,8 @@ function MeetingVote() {
           />
         </VoteTableWrapper>
       </Contents>
-      <Footer>
-        <FullHeightButtonGroup
-          fullWidth
-          disableElevation
-          variant="contained"
-          aria-label="Disabled elevation buttons"
-        >
-          <Button
-            color="secondary"
-            onClick={() => {
-              navigate(`/meetings/${meetingId}`);
-            }}
-          >
-            다음에
-          </Button>
-          <Button
-            onClick={() => {
-              handleClickVote();
-            }}
-          >
-            투표
-          </Button>
-        </FullHeightButtonGroup>
-      </Footer>
-      <InputUsernameModal
-        show={showUsernameModal}
-        usernameList={userList.map((user) => user.username)}
-        onConfirm={handleUsernameConfirm}
-        onCancel={() => setShowUsernameModal(false)}
-      />
       <ScrollDownFloatingButton onClick={handleScollDownButtonClick} show={hasMoreBottomScroll} />
+      <MeetingVoteFooter />
     </Page>
   );
 }
